inference_on_secure_cloud_run/requirements/inference.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
fastapi==0.109.0 ; python_version >= "3.11" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.11" and python_version < "4.0"
gax-google-logging-v2==0.8.3 ; python_version >= "3.11" and python_version < "4.0"
gax-google-pubsub-v1==0.8.3 ; python_version >= "3.11" and python_version < "4.0"
gcloud==0.18.3 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.15.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.26.2 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.14.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
google-gax==0.12.5 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-logging-v2==0.8.1 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-pubsub-v1==0.8.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.60.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
ply==3.8 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.35.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.26.0 ; python_version >= "3.11" and python_version < "4.0"
